move update_color_abundances to query module
